requirements.txt
flask
flask-socketio
pyperclip
requests
Pillow 